# Merges function sections