--- a/polytracker/include/polytracker/passes/taint_tracking.h
+++ b/polytracker/include/polytracker/passes/taint_tracking.h
@@ -32,6 +32,9 @@ public:
   void visitGetElementPtrInst(llvm::GetElementPtrInst &gep);
   void visitBranchInst(llvm::BranchInst &bi);
   void visitSwitchInst(llvm::SwitchInst &si);
+  void visitSelectInst(llvm::SelectInst &si);
+  void visitIndirectBrInst(llvm::IndirectBrInst &ibi);
+  void visitInvokeInst(llvm::InvokeInst &ii);
 };
 
 } // namespace polytracker