--- a/polytracker/include/polytracker/passes/tainted_control_flow.h
+++ b/polytracker/include/polytracker/passes/tainted_control_flow.h
@@ -41,9 +41,9 @@ public:
   void visitGetElementPtrInst(llvm::GetElementPtrInst &gep);
   void visitBranchInst(llvm::BranchInst &bi);
   void visitSwitchInst(llvm::SwitchInst &si);
-  void visitSelectInst(llvm::SelectInst &si);
-  void visitIndirectBrInst(llvm::IndirectBrInst &ibi);
-  void visitInvokeInst(llvm::InvokeInst &ii);
+  // void visitSelectInst(llvm::SelectInst &si);
+  // void visitIndirectBrInst(llvm::IndirectBrInst &ibi);
+  // void visitInvokeInst(llvm::InvokeInst &ii);
 
   void instrumentFunctionEnter(llvm::Function &func);
   void visitReturnInst(llvm::ReturnInst &ri);