--- a/polytracker/include/polytracker/passes/tainted_control_flow.h
+++ b/polytracker/include/polytracker/passes/tainted_control_flow.h
@@ -38,8 +38,8 @@ public:
 
   llvm::PreservedAnalyses run(llvm::Module &mod,
                               llvm::ModuleAnalysisManager &mam);
-  void visitGetElementPtrInst(llvm::GetElementPtrInst &gep);
-  // void visitBranchInst(llvm::BranchInst &bi);
+  // void visitGetElementPtrInst(llvm::GetElementPtrInst &gep);
+  void visitBranchInst(llvm::BranchInst &bi);
   // void visitSwitchInst(llvm::SwitchInst &si);
   // void visitSelectInst(llvm::SelectInst &si);
   // void visitIndirectBrInst(llvm::IndirectBrInst &ibi);