--- a/polytracker/include/taintdag/control_flow_log.h
+++ b/polytracker/include/taintdag/control_flow_log.h
@@ -1,4 +1,3 @@
-
 /*
  * Copyright (c) 2022-present, Trail of Bits, Inc.
  * All rights reserved.
@@ -9,28 +8,13 @@
 
 #pragma once
 
+#include "taintdag/control_flow_log_encoding.h"
 #include "taintdag/outputfile.h"
 #include "taintdag/section.h"
 #include "taintdag/taint.h"
 #include "taintdag/util.h"
 
 namespace taintdag {
-
-namespace detail {
-// A uint32_t varint encoded by setting highest bit for all but the final byte.
-// Requires up to 5 bytes of storage as each output byte uses 7 input bits.
-// Total maximum need is floor(32/7) = 5. Returns number of bytes required.
-size_t varint_encode(uint32_t val, uint8_t *buffer) {
-  auto orig_buffer = buffer;
-  while (val >= 0x80) {
-    *buffer++ = 0x80 | (val & 0x7f);
-    val >>= 7;
-  }
-  *buffer++ = val & 0x7f;
-  return buffer - orig_buffer;
-}
-// TODO (hbrodin): Should probably used std::span
-} // namespace detail
 
 struct ControlFlowLog : public SectionBase {
   enum EventType {
@@ -49,7 +33,7 @@ struct ControlFlowLog : public SectionBase {
   void function_event(EventType evt, uint32_t function_id) {
     uint8_t buffer[6];
     buffer[0] = static_cast<uint8_t>(evt);
-    auto used = detail::varint_encode(function_id, &buffer[1]);
+    auto used = varint_encode(function_id, &buffer[1]);
     auto total = used + 1;
 
     if (auto wctx = write(total)) {
@@ -71,9 +55,9 @@ struct ControlFlowLog : public SectionBase {
     // 1 byte event, <= 5 bytes function id, <= 5 bytes label
     uint8_t buffer[11];
     buffer[0] = static_cast<uint8_t>(TaintedControlFlow);
-    auto used = detail::varint_encode(function_id, &buffer[1]);
+    auto used = varint_encode(function_id, &buffer[1]);
     auto total = used + 1;
-    used = detail::varint_encode(label, &buffer[total]);
+    used = varint_encode(label, &buffer[total]);
     total += used;
 
     if (auto wctx = write(total)) {