… .github/workflows/build.yml
… .trunk/trunk.yaml
… CMakeLists.txt
… docs/tdag.md
… examples/Dockerfile-acropalypse.demo
… examples/Dockerfile-daedalus-nitf.demo
… examples/Dockerfile-daedalus-pdf.demo
… examples/Dockerfile-ffmpeg.demo
… examples/Dockerfile-file.demo
… examples/Dockerfile-jq.demo
… examples/Dockerfile-libjpeg.demo
… examples/Dockerfile-listgen.demo
… examples/Dockerfile-mupdf.demo
… examples/Dockerfile-nitro-nitf.demo
… examples/Dockerfile-openjpeg.demo
… examples/Dockerfile-pdfium.demo
… examples/Dockerfile-poppler.demo
… examples/Dockerfile-qpdf.demo
… examples/Dockerfile-xpdf.demo
… examples/analysis/timings.md
… examples/analysis/ubet/Dockerfile.nitro
… examples/analysis/ubet/build_nitro.sh
… examples/analysis/ubet/eval.py
… examples/analysis/ubet/eval_nitro.py
… examples/http/httpd/Dockerfile
… examples/http/picohttpparser/Dockerfile
… polytracker/include/taintdag/control_flow_log_encoding.h
… polytracker/include/taintdag/fntrace.h
… polytracker/include/taintdag/polytracker.h
… polytracker/include/taintdag/section.h